Move AzurePathError
