+++ sdk/storage/src/core/errors.rs
@@ -3,8 +3,6 @@
 pub enum Error {
     #[error(transparent)]
     CoreError(#[from] azure_core::Error),
-    #[error("Parse error: {0}")]
-    ParseError(#[from] azure_core::ParseError),
     #[error("Parsing error: {0}")]
     ParsingError(#[from] azure_core::ParsingError),
     #[error("Permission error: {0}")]
@@ -68,4 +66,17 @@ pub enum Error {
     CRC64Not8BytesLong(usize),
     #[error("At least one of these headers must be present: {0:?}")]
     HeadersNotFound(Vec<String>),
+}
+
+#[non_exhaustive]
+#[derive(Debug, thiserror::Error)]
+pub enum AzurePathParseError {
+    #[error("path separator not found")]
+    PathSeparatorNotFoundError,
+    #[error("multiple path separators found")]
+    MultiplePathSeparatorsFoundError,
+    #[error("missing container name")]
+    MissingContainerError,
+    #[error("missing blob name")]
+    MissingBlobError,
 }
